+++ app/javascript/components/Editor.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { EditorState } from 'prosemirror-state'
 import { EditorView } from 'prosemirror-view'
-import { pluginKey as commentPluginKey } from './editor-config/plugin-comment'
+import { commentPluginKey } from './editor-config/comments'
 // import applyDevTools from 'prosemirror-dev-tools'
 
 class Editor extends React.Component {
