+++ app/javascript/components/editor-config/annotation-popup.js
@@ -1,4 +1,4 @@
-import { addAnnotation } from './plugin-comment'
+import { addAnnotation } from './comments'
 
 export default {
   comments: {
